requirements.txt
pytest==8.3.5
pydantic==2.5.3
python-dotenv==1.0.1
tqdm==4.66.2
google-generativeai==0.3.2
instructor==0.6.1
